+++ repo2docker/contentproviders/git.py
@@ -21,29 +21,24 @@ class Git(ContentProvider):
 
         # make a, possibly shallow, clone of the remote repository
         try:
-            cmd = ["git", "clone", "--recursive"]
+            cmd = ["git", "clone"]
             if ref is None:
-                # check out of HEAD is performed after the clone is complete
                 cmd.extend(["--depth", "1"])
             else:
-                # don't check out HEAD, the given ref will be checked out later
-                # this prevents HEAD's submodules to be cloned if ref doesn't have them
-                cmd.extend(["--no-checkout"])
+                cmd.extend(["--branch", ref])
             cmd.extend([repo, output_dir])
             for line in execute_cmd(cmd, capture=yield_output):
                 yield line
 
         except subprocess.CalledProcessError as e:
-            msg = "Failed to clone repository from {repo}.".format(repo=repo)
+            msg = "Failed to clone repository from {repo}".format(repo=repo)
+            if ref is not None:
+                msg += " (ref {ref})".format(ref=ref)
+            msg += "."
             raise ContentProviderException(msg) from e
 
         # check out the specific ref given by the user
         if ref is not None:
-            # check out ref as it has not been done yet
-            for line in execute_cmd(
-                ["git", "checkout", ref], cwd=output_dir, capture=yield_output
-            ):
-                yield line
             hash = check_ref(ref, output_dir)
             if hash is None:
                 self.log.error(