+++ docs/source/changelog.rst
@@ -10,7 +10,6 @@ Release date: TBD
 
 New features
 ------------
-- Fix for submodule check out in :pr:`809` by :user:`davidbrochart`.
 - Add support for Figshare in :pr:`788` by :user:`nuest`.
 - Add support for Dataverse in :pr:`739` by :user:`Xarthisius`.
 
@@ -20,7 +19,7 @@ API changes
 
 Bug fixes
 ---------
-
+- Fix for submodule check out in :pr:`809` by :user:`davidbrochart`.
 
 
 Version 0.10.0